Adding an MVC approach to request management
