+++ controllers/messages.controller.js
@@ -1,0 +1,12 @@
+function getMessages(req, res) {
+  res.send("<ul><li>Hello Albert Einstein!</li></ul>");
+}
+
+function postMessage(req, res) {
+  console.log("Updating Messages...");
+}
+
+module.exports = {
+  getMessages,
+  postMessage,
+};
