+++ models/friends.model.js
@@ -1,0 +1,12 @@
+const friends = [
+  {
+    id: 0,
+    name: "Albert Einstein",
+  },
+  {
+    id: 1,
+    name: "Sir Isaac Newton",
+  },
+];
+
+module.exports = friends;
